# hellokratos/api/rider/v1/order_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v3.19.4
// source: rider/v1/order.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationOrderAcceptOrder = "/api.rider.v1.Order/AcceptOrder"
const OperationOrderCreateOrder = "/api.rider.v1.Order/CreateOrder"
const OperationOrderGetOrderDetail = "/api.rider.v1.Order/GetOrderDetail"
const OperationOrderGetOrderList = "/api.rider.v1.Order/GetOrderList"
const OperationOrderUpdateOrderStatus = "/api.rider.v1.Order/UpdateOrderStatus"

type OrderHTTPServer interface {
	// AcceptOrder 接单
	AcceptOrder(context.Context, *AcceptOrderRequest) (*AcceptOrderReply, error)
	// CreateOrder 创建订单
	CreateOrder(context.Context, *CreateOrderRequest) (*CreateOrderReply, error)
	// GetOrderDetail 获取订单详情
	GetOrderDetail(context.Context, *GetOrderDetailRequest) (*GetOrderDetailReply, error)
	// GetOrderList 获取订单列表
	GetOrderList(context.Context, *GetOrderListRequest) (*GetOrderListReply, error)
	// UpdateOrderStatus 更新订单状态
	UpdateOrderStatus(context.Context, *UpdateOrderStatusRequest) (*UpdateOrderStatusReply, error)
}

func RegisterOrderHTTPServer(s *http.Server, srv OrderHTTPServer) {
	r := s.Route("/")
	r.POST("/rider/v1/order/create", _Order_CreateOrder0_HTTP_Handler(srv))
	r.GET("/rider/v1/orders", _Order_GetOrderList0_HTTP_Handler(srv))
	r.GET("/rider/v1/order/{order_id}", _Order_GetOrderDetail0_HTTP_Handler(srv))
	r.POST("/rider/v1/order/accept", _Order_AcceptOrder0_HTTP_Handler(srv))
	r.POST("/rider/v1/order/status", _Order_UpdateOrderStatus0_HTTP_Handler(srv))
}

func _Order_CreateOrder0_HTTP_Handler(srv OrderHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateOrderRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationOrderCreateOrder)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateOrder(ctx, req.(*CreateOrderRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateOrderReply)
		return ctx.Result(200, reply)
	}
}

func _Order_GetOrderList0_HTTP_Handler(srv OrderHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetOrderListRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationOrderGetOrderList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetOrderList(ctx, req.(*GetOrderListRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetOrderListReply)
		return ctx.Result(200, reply)
	}
}

func _Order_GetOrderDetail0_HTTP_Handler(srv OrderHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetOrderDetailRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationOrderGetOrderDetail)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetOrderDetail(ctx, req.(*GetOrderDetailRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetOrderDetailReply)
		return ctx.Result(200, reply)
	}
}

func _Order_AcceptOrder0_HTTP_Handler(srv OrderHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in AcceptOrderRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationOrderAcceptOrder)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.AcceptOrder(ctx, req.(*AcceptOrderRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*AcceptOrderReply)
		return ctx.Result(200, reply)
	}
}

func _Order_UpdateOrderStatus0_HTTP_Handler(srv OrderHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateOrderStatusRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationOrderUpdateOrderStatus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateOrderStatus(ctx, req.(*UpdateOrderStatusRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateOrderStatusReply)
		return ctx.Result(200, reply)
	}
}

type OrderHTTPClient interface {
	// AcceptOrder 接单
	AcceptOrder(ctx context.Context, req *AcceptOrderRequest, opts ...http.CallOption) (rsp *AcceptOrderReply, err error)
	// CreateOrder 创建订单
	CreateOrder(ctx context.Context, req *CreateOrderRequest, opts ...http.CallOption) (rsp *CreateOrderReply, err error)
	// GetOrderDetail 获取订单详情
	GetOrderDetail(ctx context.Context, req *GetOrderDetailRequest, opts ...http.CallOption) (rsp *GetOrderDetailReply, err error)
	// GetOrderList 获取订单列表
	GetOrderList(ctx context.Context, req *GetOrderListRequest, opts ...http.CallOption) (rsp *GetOrderListReply, err error)
	// UpdateOrderStatus 更新订单状态
	UpdateOrderStatus(ctx context.Context, req *UpdateOrderStatusRequest, opts ...http.CallOption) (rsp *UpdateOrderStatusReply, err error)
}

type OrderHTTPClientImpl struct {
	cc *http.Client
}

func NewOrderHTTPClient(client *http.Client) OrderHTTPClient {
	return &OrderHTTPClientImpl{client}
}

// AcceptOrder 接单
func (c *OrderHTTPClientImpl) AcceptOrder(ctx context.Context, in *AcceptOrderRequest, opts ...http.CallOption) (*AcceptOrderReply, error) {
	var out AcceptOrderReply
	pattern := "/rider/v1/order/accept"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationOrderAcceptOrder))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// CreateOrder 创建订单
func (c *OrderHTTPClientImpl) CreateOrder(ctx context.Context, in *CreateOrderRequest, opts ...http.CallOption) (*CreateOrderReply, error) {
	var out CreateOrderReply
	pattern := "/rider/v1/order/create"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationOrderCreateOrder))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetOrderDetail 获取订单详情
func (c *OrderHTTPClientImpl) GetOrderDetail(ctx context.Context, in *GetOrderDetailRequest, opts ...http.CallOption) (*GetOrderDetailReply, error) {
	var out GetOrderDetailReply
	pattern := "/rider/v1/order/{order_id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationOrderGetOrderDetail))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetOrderList 获取订单列表
func (c *OrderHTTPClientImpl) GetOrderList(ctx context.Context, in *GetOrderListRequest, opts ...http.CallOption) (*GetOrderListReply, error) {
	var out GetOrderListReply
	pattern := "/rider/v1/orders"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationOrderGetOrderList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateOrderStatus 更新订单状态
func (c *OrderHTTPClientImpl) UpdateOrderStatus(ctx context.Context, in *UpdateOrderStatusRequest, opts ...http.CallOption) (*UpdateOrderStatusReply, error) {
	var out UpdateOrderStatusReply
	pattern := "/rider/v1/order/status"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationOrderUpdateOrderStatus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}