# hellokratos/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /rider/v1/order/accept:
        post:
            tags:
                - Order
            description: 接单
            operationId: Order_AcceptOrder
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.rider.v1.AcceptOrderRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rider.v1.AcceptOrderReply'
    /rider/v1/order/create:
        post:
            tags:
                - Order
            description: 创建订单
            operationId: Order_CreateOrder
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.rider.v1.CreateOrderRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rider.v1.CreateOrderReply'
    /rider/v1/order/status:
        post:
            tags:
                - Order
            description: 更新订单状态
            operationId: Order_UpdateOrderStatus
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.rider.v1.UpdateOrderStatusRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rider.v1.UpdateOrderStatusReply'
    /rider/v1/order/{orderId}:
        get:
            tags:
                - Order
            description: 获取订单详情
            operationId: Order_GetOrderDetail
            parameters:
                - name: orderId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rider.v1.GetOrderDetailReply'
    /rider/v1/orders:
        get:
            tags:
                - Order
            description: 获取订单列表
            operationId: Order_GetOrderList
            parameters:
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rider.v1.GetOrderListReply'
...
components:
    schemas:
        api.rider.v1.AcceptOrderReply:
            type: object
            properties:
                success:
                    type: boolean
                message:
                    type: string
            description: 接单响应
        api.rider.v1.AcceptOrderRequest:
            type: object
            properties:
                orderId:
                    type: string
                riderId:
                    type: string
            description: 接单请求
...
        api.rider.v1.CreateOrderReply:
            type: object
            properties:
                success:
                    type: boolean
                message:
                    type: string
                order:
                    $ref: '#/components/schemas/api.rider.v1.OrderInfo'
            description: 创建订单响应
        api.rider.v1.CreateOrderRequest:
            type: object
            properties:
                origin:
                    type: string
                destination:
                    type: string
                originLat:
                    type: number
                    format: float
                originLng:
                    type: number
                    format: float
                destLat:
                    type: number
                    format: float
                destLng:
                    type: number
                    format: float
                amount:
                    type: number
                    format: float
            description: 创建订单请求
...
        api.rider.v1.GetOrderDetailReply:
            type: object
            properties:
                order:
                    $ref: '#/components/schemas/api.rider.v1.OrderInfo'
            description: 获取订单详情响应
        api.rider.v1.GetOrderListReply:
            type: object
            properties:
                orders:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.rider.v1.OrderInfo'
                total:
                    type: integer
                    format: int32
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
            description: 获取订单列表响应
...
        api.rider.v1.OrderInfo:
            type: object
            properties:
                id:
                    type: string
                orderNo:
                    type: string
                status:
                    type: integer
                    format: int32
                origin:
                    type: string
                destination:
                    type: string
                originLat:
                    type: number
                    format: float
                originLng:
                    type: number
                    format: float
                destLat:
                    type: number
                    format: float
                destLng:
                    type: number
                    format: float
                distance:
                    type: number
                    format: float
                amount:
                    type: number
                    format: float
                riderId:
                    type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
            description: 订单信息
...
        api.rider.v1.UpdateOrderStatusReply:
            type: object
            properties:
                success:
                    type: boolean
                message:
                    type: string
            description: 更新订单状态响应
        api.rider.v1.UpdateOrderStatusRequest:
            type: object
            properties:
                orderId:
                    type: string
                status:
                    type: integer
                    format: int32
            description: 更新订单状态请求
...